+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 - editor
     - change keyboard shortcuts for moving widgets (now `mod + arrows`) and navigating in widgets (now `arrows`) to feel more natural with the project tree view.
 
+- widgets
+    - script: bypass keyboard shortcuts already used by the editor if it is enabled 
 
 ## 1.9.6